Separate functions into a utils.py and plot.py file
